+++ backend-springboot/src/main/java/com/nlu/service/ExamNumberService.java
@@ -4,5 +4,5 @@ import com.nlu.model.entity.ExamNumber;
 import org.springframework.http.ResponseEntity;
 
 public interface ExamNumberService {
-    public ResponseEntity<ExamNumber> getExamNumberById(Long id, Long idUser);
+    public ResponseEntity<ExamNumber> getExamNumberById(Integer id, Long idUser);
 }
